refactor base transformers
