+++ feature_engine/creation/cyclical_features.py
@@ -4,6 +4,8 @@ import numpy as np
 import pandas as pd
 from sklearn.utils.validation import check_is_fitted
 
+from feature_engine._base_transformers.base_numerical import BaseNumericalTransformer
+from feature_engine._docstrings.methods import _fit_transform_docstring
 from feature_engine._docstrings.fit_attributes import (
     _feature_names_in_docstring,
     _n_features_in_docstring,
@@ -18,7 +20,6 @@ from feature_engine._docstrings.substitute import Substitution
 from feature_engine._variable_handling.init_parameter_checks import (
     _check_init_parameter_variables,
 )
-from feature_engine.base_transformers import BaseNumericalTransformer
 
 
 @Substitution(
