+++ feature_engine/transformation/arcsin.py
@@ -21,7 +21,7 @@ from feature_engine._docstrings.substitute import Substitution
 from feature_engine._variable_handling.init_parameter_checks import (
     _check_init_parameter_variables,
 )
-from feature_engine.base_transformers import BaseNumericalTransformer
+from feature_engine._base_transformers.base_numerical import BaseNumericalTransformer
 from feature_engine.tags import _return_tags
