+++ feature_engine/transformation/yeojohnson.py
@@ -6,6 +6,8 @@ from typing import List, Optional, Union
 import pandas as pd
 import scipy.stats as stats
 
+from feature_engine._base_transformers.base_numerical import BaseNumericalTransformer
+from feature_engine._docstrings.methods import _fit_transform_docstring
 from feature_engine._docstrings.fit_attributes import (
     _feature_names_in_docstring,
     _n_features_in_docstring,
@@ -17,7 +19,6 @@ from feature_engine._docstrings.substitute import Substitution
 from feature_engine._variable_handling.init_parameter_checks import (
     _check_init_parameter_variables,
 )
-from feature_engine.base_transformers import BaseNumericalTransformer
 
 
 @Substitution(
